pyproject.toml
[tool.poetry]
name = "firecode"
version = "0.1.0"
description = ""
authors = ["verdgil <me@verdgil.org>"]

[tool.poetry.dependencies]
python = "^3.9"
djangorestframework = "^3.12.2"
django-filter = "^2.4.0"
psycopg2-binary = "^2.8.6"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
